Move asset stuff to gi folder
